detector/requirements.txt
dnspython>=2.4.0
httpx>=0.25.0
scapy>=2.5.0
